user authentication

+++ NodeJs/conFusionServer/routes/leaderRouter.js
@@ -19,28 +19,46 @@ leaderRouter.route('/')
          }, (err) => next(err))
          .catch((err) => next(err));
    })
-   .post(authenticate.verifyUser, (req, res, next) => {
-      Leaders.create(req.body)
-         .then((leader) => {
-            console.log('Leader Created ', leader);
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(leader);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Leaders.create(req.body)
+            .then((leader) => {
+               console.log('Leader Created ', leader);
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(leader);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .put(authenticate.verifyUser, (req, res, next) => {
-      res.statusCode = 403;
-      res.end('PUT operation not supported on /leaders');
+   .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         res.statusCode = 403;
+         res.end('PUT operation not supported on /leaders');
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .delete(authenticate.verifyUser, (req, res, next) => {
-      Leaders.remove({})
-         .then((resp) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(resp);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Leaders.remove({})
+            .then((resp) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(resp);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    });
 
 leaderRouter.route('/:leaderId')
@@ -53,31 +71,49 @@ leaderRouter.route('/:leaderId')
          }, (err) => next(err))
          .catch((err) => next(err));
    })
-   .post(authenticate.verifyUser, (req, res, next) => {
-      res.statusCode = 403;
-      res.end('POST operation not supported on /leaders/' + req.params.leaderId);
+   .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         res.statusCode = 403;
+         res.end('POST operation not supported on /leaders/' + req.params.leaderId);
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .put(authenticate.verifyUser, (req, res, next) => {
-      Leaders.findByIdAndUpdate(req.params.leaderId, {
-            $set: req.body
-         }, {
-            new: true
-         })
-         .then((leader) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(leader);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Leaders.findByIdAndUpdate(req.params.leaderId, {
+               $set: req.body
+            }, {
+               new: true
+            })
+            .then((leader) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(leader);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .delete(authenticate.verifyUser, (req, res, next) => {
-      Leaders.findByIdAndRemove(req.params.leaderId)
-         .then((resp) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(resp);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Leaders.findByIdAndRemove(req.params.leaderId)
+            .then((resp) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(resp);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    });
 
 module.exports = leaderRouter;
