+++ NodeJs/conFusionServer/routes/promoRouter.js
@@ -19,28 +19,46 @@ promotionRouter.route('/')
          }, (err) => next(err))
          .catch((err) => next(err));
    })
-   .post(authenticate.verifyUser, (req, res, next) => {
-      Promotions.create(req.body)
-         .then((promotion) => {
-            console.log('Promotion Created ', promotion);
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(promotion);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Promotions.create(req.body)
+            .then((promotion) => {
+               console.log('Promotion Created ', promotion);
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(promotion);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .put(authenticate.verifyUser, (req, res, next) => {
-      res.statusCode = 403;
-      res.end('PUT operation not supported on /promotions');
+   .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         res.statusCode = 403;
+         res.end('PUT operation not supported on /promotions');
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .delete(authenticate.verifyUser, (req, res, next) => {
-      Promotions.remove({})
-         .then((resp) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(resp);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Promotions.remove({})
+            .then((resp) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(resp);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    });
 
 promotionRouter.route('/:promotionId')
@@ -53,31 +71,49 @@ promotionRouter.route('/:promotionId')
          }, (err) => next(err))
          .catch((err) => next(err));
    })
-   .post(authenticate.verifyUser, (req, res, next) => {
-      res.statusCode = 403;
-      res.end('POST operation not supported on /promotions/' + req.params.promotionId);
+   .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         res.statusCode = 403;
+         res.end('POST operation not supported on /promotions/' + req.params.promotionId);
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .put(authenticate.verifyUser, (req, res, next) => {
-      Promotions.findByIdAndUpdate(req.params.promotionId, {
-            $set: req.body
-         }, {
-            new: true
-         })
-         .then((promotion) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(promotion);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Promotions.findByIdAndUpdate(req.params.promotionId, {
+               $set: req.body
+            }, {
+               new: true
+            })
+            .then((promotion) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(promotion);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    })
-   .delete(authenticate.verifyUser, (req, res, next) => {
-      Promotions.findByIdAndRemove(req.params.promotionId)
-         .then((resp) => {
-            res.statusCode = 200;
-            res.setHeader('Content-Type', 'application/json');
-            res.json(resp);
-         }, (err) => next(err))
-         .catch((err) => next(err));
+   .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+      if (req.user.admin) {
+         Promotions.findByIdAndRemove(req.params.promotionId)
+            .then((resp) => {
+               res.statusCode = 200;
+               res.setHeader('Content-Type', 'application/json');
+               res.json(resp);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+      } else {
+         var err = new Error('You are not authorized to perform this operation!');
+         err.status = 403;
+         next(err);
+      }
    });
 
 module.exports = promotionRouter;
